+++ quantum_backend_bench/utils/__init__.py
@@ -1,1 +1,5 @@
 """Utility helpers."""
+
+from quantum_backend_bench.utils import notebook
+
+__all__ = ["notebook"]